CIV-9201 increase coverage
